fix: reduce batch_size and cast ref_model to bf16 to fit in 16GB
Chunked log_softmax alone wasn't enough - OOM'd again on step 1
(allocation size dropped 3.37->1.86 GiB, but baseline usage was still
13.77/15.47 GiB). batch_size 4->2 (cheaper to cut than group_size,
which determines advantage-estimate quality per prompt) and ref_model
now permanently bf16 (inference-only, no gradients, no precision cost
beyond what the policy already accepts via autocast).

+++ src/trainer/grpo.py
@@ -156,7 +156,10 @@ class GRPOTrainer(BaseTrainer):
 
     def model_post_init(self, __context):
         self.model.to(self.config.device)
-        self.ref_model.to(self.config.device)
+        # bf16, not fp32: ref_model only ever does a forward pass (no gradients, no
+        # optimizer state), so this halves its memory with no precision cost beyond
+        # what the policy's own forward pass already accepts via autocast bf16.
+        self.ref_model.to(self.config.device, dtype=torch.bfloat16)
         self.ref_model.eval()
         self.ref_model.requires_grad_(False)
